sdk: Split attachment tests in separate file
