+++ crates/matrix-sdk/tests/integration/room/mod.rs
@@ -1,3 +1,4 @@
+mod attachment;
 mod common;
 mod joined;
 mod left;
